svn merge -r9875:9886 https://eris.llnl.gov/svn/slurm/branches/slurm-1.1

+++ trunk/src/plugins/select/bluegene/plugin/select_bluegene.c
@@ -428,7 +428,7 @@ extern int select_p_get_info_from_plugin (enum select_data_info info,
 	return SLURM_SUCCESS;
 }
 
-extern int select_p_update_node_state (int index, int state)
+extern int select_p_update_node_state (int index, uint16_t state)
 {
 	int x;
 #ifdef HAVE_BG
