+++ trunk/src/plugins/select/linear/select_linear.c
@@ -768,7 +768,7 @@ extern int select_p_get_info_from_plugin (enum select_data_info info,
 	return SLURM_SUCCESS;
 }
 
-extern int select_p_update_node_state (int index, int state)
+extern int select_p_update_node_state (int index, uint16_t state)
 {
 	return SLURM_SUCCESS;
 }
